Revert "rebase > merge (#14)"
This reverts commit 3108b42c7116dc23944f2c5416dbf507feb1efd0.
