--- Hip-Pal_v2/requirements.txt
@@ -1,2 +1,0 @@
-monai-weekly[gdown, nibabel, tqdm, ignite]
-matplotlib
